requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
aiohttp==3.9.1
mcp==1.0.0
python-multipart==0.0.6
websockets==12.0
azure-cognitiveservices-speech==1.36.0
SpeechRecognition==3.10.0
google-api-python-client==2.109.0
google-auth-oauthlib==1.1.0